Getting Started on Battlemode
Battlemode (Containters) is going to be a combat oriented TTRPG with a tactical combat focus and light attention to the RPG side. This is right now basically a clone from my unnamed project, sans all the pages and content. Just trying to get the Docsify up and running for now.